update lib version
